Massive module layout overhaul
