+++ dwm1001/exceptions.py
@@ -1,0 +1,11 @@
+# Local module defined exceptions
+class ParsingError(Exception):
+    """! Exception raised when parsing fails for shell command output."""
+
+    pass
+
+
+class ReservedGPIOPinError(Exception):
+    """! Exception raised when trying to use a reserved GPIO pin."""
+
+    pass
