+++ tests/test_TagPosition.py
@@ -6,7 +6,7 @@ import pytest
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
 # Import modules under test
-from dwm1001 import TagPosition, ParsingError
+from dwm1001.dwm1001 import TagPosition, ParsingError
 
 
 # ************************* Begin Tests ************************* #
