+++ tests/test_UartDwm1001.py
@@ -9,7 +9,7 @@ from unittest import mock
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
 # Import modules under test
-from dwm1001 import UartDwm1001, AccelerometerData, NodeMode
+from dwm1001.dwm1001 import UartDwm1001, AccelerometerData, NodeMode
 
 
 # ************************* Mock Serial ************************* #
